requirements.txt
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.12.4